Fetch quacks new issue relationships unresolved for feed
