+++ app/(wrap)/home/page.tsx
@@ -1,6 +1,6 @@
 import Form from "../../ui/feed/form";
-import QuackWrapper from "../../ui/feed/quacks";
-import Quacks from "../../ui/feed/quacks";
+import QuackWrapper from "../../ui/feed/quackWrapper";
+import Quacks from "../../ui/feed/quackWrapper";
 import React from "react";
 
 export default async function Page() {
